dm: PTM: Check PTM root has more than one child
	Add one more sanity check: If the root port has more than
	one child, we won't enable PTM on the guest.
	This is not necessarily an error.  We flag it as
	error just because we don't have this type of hw
	configuration at development time thus this configuration
	is not tested.

Tracked-On: #5915
Acked-by: Yu Wang <yu1.wang@intel.com>
